yatube/requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
Django==2.2.16
django-behaviors==0.5.1
Faker==12.0.1
idna==3.4
iniconfig==2.0.0
mixer==7.1.2
packaging==23.0
Pillow==8.3.1
pluggy==0.13.1
py==1.11.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.14
